Added Nagios formated Output of DMR Master connection
A new config variable called NagiosStatusFile was introduced in section
General. Output formated to be used directly in Nagios is generated,
preferable to be stored in a ramdisk like /dev/shm/... .
If the program is closed, there should be a last write to this status
file with status "Unkown", but unfortunately this is not working yet.