Attempt to solve one cause of "No reply from modem" loop
"err 101" (network unreachable - effectively cant get to the master because we've lost the network)

This needs trapping and handling properly to smooth the way it currently tries to send from the socket as quickly as it can (effectively every millisecond) until the network is restored,  then does the same for "Opening IPSC" every millisecond or less until the login either times out (in which case I've edited the code to tell it to retry!) or logs into the master successfully.

This error also causes the "no reply from modem loop, however this is now trapped in all cases if the firmware is read successfully by stopping the inactivity timer, allowing the software to continue reading the modem.

It works for this particular error, but others may need to test!