Add VERSION to STDERR and log messages that output the string MMDVMHost
They now output the string MMDVMHost-yyymmdd

a) for completeness 

and;

b) to quickly output the version number by calling the host executable with no arguments so Kim can easily grab it for his dashboard.

I did think about a -v|--version argument, but this was much simpler for the purpose.