LCDproc: Make dimming the status screen optional
If you run another LCDproc client that doesn't dim the display when it displays its info, the display will effectively flash on and off (or dim/bright depending on your LCDd configuration) as it switches screens between MMDVMHost and the other client(s).

Making the dimming optional and turning it off in the host stops this "annoyance".