Updated download URL for the DMRid database
Access to the real time database feed has been discontinued. A nightly snapshot is available instead. 

http://dmr-marc.net/cgi-bin/trbo-database/

Corrected path for the right URL. Beware, the above page has an error, the URLs are not valid due to a certificate mismatch. Adding "www" to the URLs works and correct the isue nevertheless.